Implemented support for MvxPagePresentationHint in the MvxAppCompatViewPresenter. Removed logic to activate tabs in the ShowViewPagerFragment method.
